move TempDirectory to src/Common
